File icon lookup mechanism refactored
Motivation:
The old implementation was very hard to extend
and to be understood.
Because of that, a configuration file has
been introduced, which can be changed extended
with new file associations easily.
The lookup mechanism goes linearily, thus a simple
ranking system has been introduced implicitly.

The associations can be given in two ways currently:
- RegexAssociation (regular expression pattern)
- TypeAssociation (file type)